+++ crates/core/executor/src/memory.rs
@@ -1,3 +1,4 @@
+use crate::register::NUM_REGISTERS;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use vec_map::VecMap;
 
@@ -47,10 +48,8 @@ pub struct PagedMemory<V: Copy> {
 impl<V: Copy> PagedMemory<V> {
     /// The number of lower bits to ignore, since addresses (except registers) are a multiple of 4.
     const NUM_IGNORED_LOWER_BITS: usize = 2;
-    /// The number of registers in the virtual machine.
-    const NUM_REGISTERS: usize = 36;
     /// The offset subtracted from the main address space to make it contiguous.
-    const ADDR_COMPRESS_OFFSET: usize = Self::NUM_REGISTERS;
+    const ADDR_COMPRESS_OFFSET: usize = NUM_REGISTERS;
 
     /// Create a `PagedMemory` with capacity `MAX_PAGE_COUNT`.
     pub fn new_preallocated() -> Self {
@@ -155,7 +154,7 @@ impl<V: Copy> PagedMemory<V> {
     #[inline]
     const fn compress_addr(addr: u32) -> usize {
         let addr = addr as usize;
-        if addr < Self::NUM_REGISTERS {
+        if addr < NUM_REGISTERS {
             addr
         } else {
             (addr >> Self::NUM_IGNORED_LOWER_BITS) + Self::ADDR_COMPRESS_OFFSET
@@ -165,7 +164,7 @@ impl<V: Copy> PagedMemory<V> {
     /// Decompress an address from a contiguous space to the sparse address space.
     #[inline]
     const fn decompress_addr(addr: usize) -> u32 {
-        if addr < Self::NUM_REGISTERS {
+        if addr < NUM_REGISTERS {
             addr as u32
         } else {
             ((addr - Self::ADDR_COMPRESS_OFFSET) << Self::NUM_IGNORED_LOWER_BITS) as u32
