+++ crates/core/executor/src/syscalls/precompiles/sys_linux/sysclone.rs
@@ -20,10 +20,7 @@ impl Syscall for SysCloneSyscall {
     ) -> Option<u32> {
         let start_clk = rt.clk;
         let v0 = 1; // Simulate a successful clone operation
-        let read_records = Vec::new();
-        let mut write_records = Vec::new();
         let a3_record = rt.mw(Register::A3 as u32, 0);
-        write_records.push(a3_record);
         let shard = rt.current_shard();
         let event = PrecompileEvent::Linux(LinuxEvent {
             shard,
@@ -32,8 +29,8 @@ impl Syscall for SysCloneSyscall {
             a1,
             v0,
             syscall_code: syscall_code.syscall_id(),
-            read_records,
-            write_records,
+            read_records: vec![],
+            write_records: vec![a3_record],
             local_mem_access: rt.postprocess(),
         });
         let syscall_event =
