+++ crates/core/executor/src/syscalls/precompiles/sys_linux/sysread.rs
@@ -21,8 +21,6 @@ impl Syscall for SysReadSyscall {
         a1: u32,
     ) -> Option<u32> {
         let start_clk = rt.clk;
-        let read_records = Vec::new();
-        let mut write_records = Vec::new();
         let fd = a0;
         let mut v0 = 0;
         let a3_record = if fd != FD_STDIN {
@@ -31,7 +29,7 @@ impl Syscall for SysReadSyscall {
         } else {
             rt.mw(Register::A3 as u32, 0)
         };
-        write_records.push(a3_record);
+
         let shard = rt.current_shard();
         let event = PrecompileEvent::Linux(LinuxEvent {
             shard,
@@ -40,8 +38,8 @@ impl Syscall for SysReadSyscall {
             a1,
             v0,
             syscall_code: syscall_code.syscall_id(),
-            read_records,
-            write_records,
+            read_records: vec![],
+            write_records: vec![a3_record],
             local_mem_access: rt.postprocess(),
         });
         let syscall_event =
